Add money incrementation

+++ src/main/java/Player.java
@@ -32,7 +32,7 @@ public class Player {
         return money;
     }
 
-    public void setMoney(int money) {
+    public void setMoney(float money) {
         this.money = money;
     }
